Add PDF splitter sample
This sample is a Python CLI that uses Document AI's splitter
processor to determine how a PDF document should be split. This
information is then used to split the document into subdocuments
that are saved locally as PDFs. The sample does not alter the
input PDF.

There are several flags to configure the CLI:

  --input/-i path to the PDF file to be split
  --output is the path to the dir to save split PDF files
           Defaults to current directory
  --project-id is the project ID used to call the Document AI API
               Defaults to the project indicated by application
               default credentials
  --multi-region-location indicates which region to use
                          defaults to "us"
  --split-processor-type indicates what processor type to use
                         defaults to "DOCUMENT_SPLIT_PROCESSOR"

This sample will search to see if a processor of the specified
type is already present in the specified project and location.
If the processor already exists, it is used. If it does not exist
a processor of the specified type will be created in the specified
projects and location.

Unit tests mock Document AI API responses and use the file
`multi_document.pdf` to check that the PDF is split as expected.

The sample depends on the Document AI Python client library to
make Document AI AI calls and the Pike PDF library to read and
write PDFs locally.