Add PDF annotator sample (#3)
This sample is a Python CLI that uses Document AI's form parser to annotate a
PDF document. PDF annotations are added based on the key-value pairs retrivied
from Document AI form entities indicating what information is present in a
document. For instance an annotation is added for the "balance due" for an
invoice and specifies the amount. The annotates PDF is saved locally seperately
from the input PDF. The sample does not alter the input PDF.

There are several flags to configure the CLI:

  --input/-i path to the PDF file to be annotated
  --output is the path to the save the annotated PDF file
           Defaults to the input path with the suffix "_annotated"
  --project-id is the project ID used to call the Document AI API
               Defaults to the application default credentials project
  --multi-region-location indicates which region to use
                          defaults to "us"
  --split-processor-type indicates what processor type to use
                         defaults to "FORM_PARSER_PROCESSOR"

This sample will search to see if a processor of the specified type is already
present in the specified project and location. If the processor already exists,
it is used. If it does not exist a processor of the specified type will be
created in the specified projects and location.

The sample depends on the Document AI Python client library to make Document AI
AI calls and the Pike PDF library to read and write PDFs locally.