Fix splitter test and update deps (#7)
Fix splitter test and update deps

This PR changes the PDF Python splitter tests to mock the DocumentProcessorServiceClient client itself instead of the DocumentProcessorServiceClient methods used by `main.py`. This change was made to prevent the DocumentProcessorServiceClient init function from looking for application default credentials.  Before this change the DocumentProcessorServiceClient would look for credentials in the local environment which (when not present) would cause the tests to fail.

This change also updates the PyPI dependencies:

* google-cloud-documentai 1.2.1 --> 1.3.0
* pikepdf 4.3.1 --> 5.0.1 

and drops Python 3.6 tests in CI because a dependency (pikepdf) doesn't support Python 3.6 in the latest (5.0.1) and previously used (4.3.1) versions which caused these Python 3.6 tests to fail.